Convert example-middleware-analyzer to a minimal scaffold example
