--- cmd/example-middleware-analyzer/interfaces/constants.go
@@ -1,17 +1,0 @@
-// Copyright 2023 Symbl.ai SDK contributors. All Rights Reserved.
-// SPDX-License-Identifier: Apache License 2.0
-
-package interfaces
-
-const (
-	// app specific message type
-	AppSpecificMessageTypeHistorical string = "historical"
-
-	// user/app level association type
-	UserAssociationTypeTopic   string = "topic_association"
-	UserAssociationTypeTracker string = "tracker_association"
-	UserAssociationTypeEntity  string = "entity_association"
-
-	// app specific message type
-	MessageNotFound string = "**MESSAGE NOT FOUND**"
-)
