+++ cmd/example-middleware-plugin/interfaces/constants.go
@@ -1,0 +1,18 @@
+// Copyright 2023 Symbl.ai SDK contributors. All Rights Reserved.
+// SPDX-License-Identifier: Apache License 2.0
+
+package interfaces
+
+const (
+	// app specific message type
+	AppSpecificMessageTypeScaffold string = "scaffold"
+
+	// user/app level scaffold type
+	UserScaffoldTypeInsight string = "scaffold_insight"
+	UserScaffoldTypeTopic   string = "scaffold_topic"
+	UserScaffoldTypeTracker string = "scaffold_tracker"
+	UserScaffoldTypeEntity  string = "scaffold_entity"
+
+	// app specific message type
+	MessageNotFound string = "**MESSAGE NOT FOUND**"
+)
