+++ cmd/example-middleware-plugin/server/constants.go
