+++ go.mod
@@ -4,7 +4,7 @@ go 1.18
 
 require (
 	github.com/dvonthenen/rabbitmq-manager v0.1.1
-	github.com/dvonthenen/symbl-go-sdk v0.1.5-0.20230302195324-986cb6f026d8
+	github.com/dvonthenen/symbl-go-sdk v0.1.5
 	github.com/dvonthenen/websocketproxy v0.1.0-dyv.1
 	github.com/hokaccha/go-prettyjson v0.0.0-20211117102719-0474bc63780f
 	github.com/neo4j/neo4j-go-driver/v5 v5.3.0
