+++ pkg/middleware-plugin-sdk/router/entityhandler.go
