+++ pkg/middleware-plugin-sdk/router/insighthandler.go
